+++ api/v1alpha1/paas_types.go
@@ -14,8 +14,8 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/belastingdienst/opr-paas/v3/internal/fields"
-	"github.com/belastingdienst/opr-paas/v3/internal/groups"
+	"github.com/belastingdienst/opr-paas/v3/pkg/fields"
+	"github.com/belastingdienst/opr-paas/v3/pkg/groups"
 	paasquota "github.com/belastingdienst/opr-paas/v3/pkg/quota"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
