+++ internal/webhook/v1alpha1/paasconfig_webhook.go
@@ -13,8 +13,7 @@ import (
 
 	"github.com/belastingdienst/opr-paas/v3/api/v1alpha1"
 	"github.com/belastingdienst/opr-paas/v3/internal/logging"
-	"github.com/belastingdienst/opr-paas/v3/internal/templating"
-	"github.com/belastingdienst/opr-paas/v3/internal/validate"
+	"github.com/belastingdienst/opr-paas/v3/pkg/templating"
 	k8sv1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	resourcev1 "k8s.io/apimachinery/pkg/api/resource"
@@ -430,7 +429,7 @@ func validateConfigCustomField(
 
 	if customfield.Validation != "" {
 		// Must have compilable regex
-		if valid, err := validate.StringIsRegex(customfield.Validation); !valid {
+		if _, err := regexp.Compile(customfield.Validation); err != nil {
 			allErrs = append(allErrs, field.Invalid(
 				childPath.Child("validation"),
 				name,
