+++ pkg/fields/elementlist_test.go
@@ -3,7 +3,7 @@ package fields_test
 import (
 	"testing"
 
-	"github.com/belastingdienst/opr-paas/v3/internal/fields"
+	"github.com/belastingdienst/opr-paas/v3/pkg/fields"
 	"github.com/stretchr/testify/assert"
 )
