+++ pkg/templating/main.go
@@ -10,7 +10,7 @@ import (
 	"github.com/belastingdienst/opr-paas/v3/api"
 	"github.com/belastingdienst/opr-paas/v3/api/v1alpha1"
 	"github.com/belastingdienst/opr-paas/v3/api/v1alpha2"
-	"github.com/belastingdienst/opr-paas/v3/internal/fields"
+	"github.com/belastingdienst/opr-paas/v3/pkg/fields"
 )
 
 // PaasUnion is an interface representing either a v1alpha1.Paas or a v1alpha2.Paas
