+++ pkg/templating/yaml_to.go
@@ -1,7 +1,7 @@
 package templating
 
 import (
-	"github.com/belastingdienst/opr-paas/v3/internal/fields"
+	"github.com/belastingdienst/opr-paas/v3/pkg/fields"
 	"github.com/goccy/go-yaml"
 )
